move all of the npm package config to a module
